pyproject.toml
[project]
name = "lab"
version = "0.1.0"
description = "Laboratory experiment tracking system — data model"
requires-python = ">=3.11"
dependencies = [
    "sqlmodel>=0.0.22",
    "sqlalchemy>=2.0.36",
    "psycopg[binary]>=3.2",
    "alembic>=1.14",
    "pydantic-settings>=2.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lab"]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "faker>=40.15.0",
    "pytest>=8",
    "pytest-cov>=5",
    "pytest-spec>=5",
    "ruff>=0.7",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--spec --cov=src/lab --cov-branch --cov-fail-under=100 --cov-report=term-missing"

[tool.coverage.run]
branch = true
source = ["src/lab"]

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 100
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
